[enhance](cloud)support SHOW-TRANSACTION
